Some reasonable baselines run
Simple baselines both regressing and classifying the outcome data.
The takeaway is that classifying an average over time of the
cqt (i.e. a vector of how much each frequency has been heard) is
best done with a logistic regression. Errors are largely from
misclassifying majors as minors (i.e. A maj classified as A minor)