Include flat test loss
We decided that it would be important to see how the test loss
changes without all the class imbalance issues. This is because,
we think, the class imbalance makes the real test loss dart about.
Seeing as we don't know the test set, we decided to train on a
flat set of keys.